refactor(google_genai): move Gemma-specific helpers into gemma.dart
Mirrors the imagen.dart layout (PR #258) so model-family logic lives
in its own file. The generateContent path is still shared with Gemini
via gemmaToGeminiOptions + the existing toGeminiSettings helper, so
the Gemma branch in createModel stays put — only the data (info
constants, predicates, reasoning-strip filter, options mapper) moves.

Renamed _gemmaToGeminiOptions to gemmaToGeminiOptions since it is
now consumed across files. Tests now import from src/gemma.dart.

Sets up the file-per-family pattern for upcoming model families
(veo, lyria, aqa, nano-banana-pro, deep-research).
