+++ packages/genkit_google_genai/lib/src/common_plugin.dart
@@ -21,6 +21,7 @@ import 'package:schemantic/schemantic.dart';
 
 import 'aggregation.dart';
 import 'api_client.dart';
+import 'gemma.dart';
 import 'generated/generativelanguage.dart' as gcl;
 import 'model.dart';
 
@@ -36,26 +37,6 @@ final commonModelInfo = ModelInfo(
     'constrained': true,
   },
 );
-
-final commonGemmaModelInfo = ModelInfo(
-  supports: {
-    'multiturn': true,
-    'media': true,
-    'tools': true,
-    'toolChoice': true,
-    'systemRole': true,
-    'constrained': 'no-tools',
-  },
-);
-
-final gemma3ModelInfo = ModelInfo(
-  supports: {...?commonGemmaModelInfo.supports, 'systemRole': false},
-);
-
-bool isGemmaModelName(String name) => name.startsWith('gemma-');
-
-bool isGemma3ModelName(String name) =>
-    name.startsWith('gemma-3-') || name.startsWith('gemma-3n-');
 
 abstract class CommonGoogleGenPlugin extends GenkitPlugin {
   Future<GenerativeLanguageBaseClient> getApiClient([String? requestApiKey]);
@@ -104,7 +85,7 @@ abstract class CommonGoogleGenPlugin extends GenkitPlugin {
             final gemmaOptions = req.config == null
                 ? GemmaOptions()
                 : GemmaOptions.$schema.parse(req.config!);
-            options = _gemmaToGeminiOptions(gemmaOptions);
+            options = gemmaToGeminiOptions(gemmaOptions);
           } else {
             options = req.config == null
                 ? GeminiOptions()
@@ -470,51 +451,6 @@ List<gcl.Content> toGeminiContent(List<Message> messages) {
         ),
       )
       .toList();
-}
-
-@visibleForTesting
-List<Message> stripReasoningParts(List<Message> messages) {
-  return messages
-      .map(
-        (m) => Message(
-          role: m.role,
-          content: m.content
-              .where(
-                (p) =>
-                    !p.isReasoning && p.metadata?['thoughtSignature'] == null,
-              )
-              .toList(),
-          metadata: m.metadata,
-        ),
-      )
-      .where((m) => m.content.isNotEmpty)
-      .toList();
-}
-
-GeminiOptions _gemmaToGeminiOptions(GemmaOptions o) {
-  return GeminiOptions(
-    apiKey: o.apiKey,
-    safetySettings: o.safetySettings,
-    codeExecution: o.codeExecution,
-    functionCallingConfig: o.functionCallingConfig,
-    thinkingConfig: o.thinkingConfig,
-    responseModalities: o.responseModalities,
-    googleSearch: o.googleSearch,
-    fileSearch: o.fileSearch,
-    temperature: o.temperature,
-    topP: o.topP,
-    topK: o.topK,
-    candidateCount: o.candidateCount,
-    stopSequences: o.stopSequences,
-    maxOutputTokens: o.maxOutputTokens,
-    responseMimeType: o.responseMimeType,
-    responseLogprobs: o.responseLogprobs,
-    logprobs: o.logprobs,
-    presencePenalty: o.presencePenalty,
-    frequencyPenalty: o.frequencyPenalty,
-    seed: o.seed,
-    speechConfig: o.speechConfig,
-  );
 }
 
 @visibleForTesting
